Chatbot/Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings==2.1.0
httpx>=0.27.0

